Add prelim TempDir struct for testing
This commit adds preliminary `TempDir` struct that can be used to
autocreate a collision-free temporary directory in the host's
temp dir location (WASI will need a seperate consideration so it's
currently not supported; I will rectify this in a subsequent commit).
The `TempDir` is based around `tempfile::TempDir` struct and similarly,
on `drop`, will autoclean itself.